particlize is now a thing. All apps run their config through root particle.js
