+++ apps/grav/webpack.grav.dev.js
@@ -1,9 +1,6 @@
 /**
  * Grav-specific webpack config.
- * This is merged over top of webpack.particle.prod.js
  *
- * Import shared PROD settings because webpack.particle.prod extracts styles to its own standalone
- * CSS files. So grav dev uses shared PROD settings, in DEV MODE to rapidly build all assets.
  */
 
 // Library Imports
@@ -11,7 +8,7 @@ const merge = require('webpack-merge');
 const RunScriptAfterEmit = require('../../tools/webpack/run-script-after-emit');
 
 // Custom Imports
-const particle = require('../../webpack.particle.prod');
+const particle = require('../../particle');
 const grav = require('./webpack.grav.shared');
 
 // Webpack Entry Points
