+++ package.json
@@ -20,14 +20,14 @@
     "start": "npm run dev:pl",
     "dev:pl": "NODE_ENV='development' webpack-dev-server --config ./apps/pl/webpack.pl.js --hot",
     "build:pl": "NODE_ENV='production' webpack --config ./apps/pl/webpack.pl.js",
-    "dev:drupal": "NODE_ENV='development' webpack --watch --progress --config ./apps/drupal/webpack.drupal.dev.js",
-    "build:drupal": "NODE_ENV='production' webpack --config ./apps/drupal/webpack.drupal.prod.js",
+    "dev:drupal": "NODE_ENV='development' webpack --watch --config ./apps/drupal/webpack.drupal.js",
+    "build:drupal": "NODE_ENV='production' webpack --config ./apps/drupal/webpack.drupal.js",
     "dev:grav": "NODE_ENV='development' webpack --watch --progress --config ./apps/grav/webpack.grav.dev.js",
     "build:grav": "NODE_ENV='production' webpack --config ./apps/grav/webpack.grav.prod.js",
     "lint:js": "eslint ./",
     "lint:scss": "stylelint '**/*.{css,scss}'",
     "lint": "npm run lint:js; npm run lint:scss",
-    "fmt:js": "npm run lint:js -- --fix && prettier --loglevel warn --write '**/*.{js,json}'",
+    "fmt:js": "npm run lint:js -- --fix && prettier --write '**/*.{js,json}'",
     "fmt:scss": "npm run lint:scss -- --fix && prettier --loglevel warn --write '**/*.{css,scss}'",
     "fmt": "npm run fmt:js; npm run fmt:scss",
     "test:pa11y": "node tools/tests/accessibility/pa11y.js",
