+++ source/_patterns/02-molecules/card/_card.scss
@@ -13,3 +13,6 @@ $card-border-width: 1px;
 .border-none {
   border-style: none;
 }
+body {
+  color: green;
+}
